test: GetChangedFiles test
Adds testing for the GetChangedFiles function. More importantly we have
found a way to test all these functions. This proved difficult because
it relies on testing git history but we dont want to effect the repos
history.

What we have ended up doing is create a git submodule directory, testing in
there and clean it up after we are done.